[CSHW-6]: Добавлен глобальный индекс для поиска

+++ src/Ozon.Route256.Practice.OrdersService/Repository/Dto/AddressDto.cs
@@ -10,6 +10,6 @@ public sealed record AddressDto
     string Street,
     string Building,
     string Apartment,
-    double Latitude,
-    double Longitude
+    decimal Latitude,
+    decimal Longitude
 );
